Make the annotated code work on 3.8, merge it into __init__ from extras.
