+++ src/ducktools/classbuilder/__init__.pyi
@@ -100,6 +100,15 @@ def make_slot_gatherer(field_type: type[Field] = Field) -> Callable[[type], dict
 def slot_gatherer(cls: type) -> dict[str, Field]:
     ...
 
+def is_classvar(hint: object) -> bool: ...
+
+def make_annotation_gatherer(
+    field_type: type[Field] = Field,
+    leave_default_values: bool = True,
+) -> Callable[[type], dict[str, Field]]: ...
+
+def annotation_gatherer(cls: type) -> dict[str, Field]: ...
+
 @typing.overload
 def slotclass(
     cls: type[_T],
@@ -116,6 +125,22 @@ def slotclass(
     *,
     methods: frozenset[MethodMaker] | set[MethodMaker] = default_methods,
     syntax_check: bool = True
+) -> Callable[[type[_T]], type[_T]]: ...
+
+@typing.overload
+def annotationclass(
+    cls: type[_T],
+    /,
+    *,
+    methods: frozenset[MethodMaker] | set[MethodMaker] = default_methods,
+) -> type[_T]: ...
+
+@typing.overload
+def annotationclass(
+        cls: None = None,
+        /,
+        *,
+        methods: frozenset[MethodMaker] | set[MethodMaker] = default_methods,
 ) -> Callable[[type[_T]], type[_T]]: ...
 
 def fieldclass(cls: type[_T]) -> type[_T]: ...
