+++ src/ducktools/classbuilder/prefab.pyi
@@ -26,8 +26,6 @@ KW_ONLY: _KW_ONLY_TYPE
 
 class PrefabError(Exception): ...
 
-def _is_classvar(hint: type | str) -> bool: ...
-
 def get_attributes(cls: type) -> dict[str, Attribute]: ...
 
 def get_init_maker(*, init_name: str="__init__") -> MethodMaker: ...
@@ -37,10 +35,6 @@ def get_repr_maker(*, recursion_safe: bool = False) -> MethodMaker: ...
 def get_eq_maker() -> MethodMaker: ...
 
 def get_iter_maker() -> MethodMaker: ...
-
-def get_frozen_setattr_maker() -> MethodMaker: ...
-
-def get_frozen_delattr_maker() -> MethodMaker: ...
 
 def get_asdict_maker() -> MethodMaker: ...
